backend/feedback-scoring-service/requirements.txt
# API Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.5.2
python-multipart>=0.0.6

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
torch>=2.0.0
transformers==4.37.2
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.25.1

# NLP Libraries
spacy==3.7.2
nltk>=3.8.1
textblob==0.17.1
regex>=2023.10.3

# spaCy dependencies
spacy-legacy==3.0.12
spacy-loggers==1.0.5
cymem==2.0.11
preshed==3.0.9
thinc==8.2.5
blis==0.7.11
murmurhash==1.0.12
wasabi==1.1.3
srsly==2.5.1
catalogue==2.0.10
typer==0.9.0
pathy>=0.10.2
smart-open>=5.2.1
confection>=0.1.3
langcodes>=3.3.0

# Database
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.1
aiohttp==3.9.3
tqdm==4.66.2
joblib==1.3.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 